--- examples/cms-dotcms/components/dotCMS/storyblock/Marks.js
@@ -1,12 +1,0 @@
-import DotLink from '../../link'
-
-export const Bold = ({ children }) => <strong>{children}</strong>
-export const Italic = ({ children }) => <em>{children}</em>
-export const Link = ({ attrs: { href, target }, children }) => (
-  <DotLink href={href} target={target}>
-    {children}
-  </DotLink>
-)
-
-export const Strike = ({ children }) => <s>{children}</s>
-export const Underline = ({ children }) => <u>{children}</u>
