--- examples/cms-dotcms/components/dotCMS/storyblock/index.js
@@ -1,4 +1,0 @@
-// Dependencies
-import dynamic from 'next/dynamic'
-
-export const DotSBRender = dynamic(() => import('./DotSBRender'))
